Moved unneccessary pages to _examples
